create cloud provider arg
